Built parallel methods for testing, probably moving to empy for workflows and jinja2 for global
Bug report confirms issue with empy integration, but the issue appears to only be raised when using empy for low level modules in the global.cylc file.
The error is suppressed when os.environ is only used in flow.cylc files and jinja2 templates the global.cylc file.  Future releases are planned
to fix the overall integration, but this can be worked around for now fairly easily.  Need to update bug report with the additional observation.