Further sketching of GSI workflow and integration to Cylc
Tasks are sketched but this has introduced errors in the Cylc validation
of the Cylc graph -- the error message is non-informative because the
same graph validates without the task information.  Need to closely
work through the task definitions / hierarchy to fix issues and
for smoothing with driving scripts.  Also need to compile WRFDA
in the common software stack for testing.